refactor(settings): de-monolith SettingsView 2273→88 (redo of deferred #192)
Split the SettingsView monolith into a thin shell plus per-tab views and
shared components, computed against CURRENT master so none of the recent
settings improvements regress (#187 chip badges, #188 glass CTAs, #190
a11y/Dynamic Type, #191 tokens, #193 EmptyStateCard notices, #196 Grok).

Shell SettingsView.swift (88 lines) now owns only the TabView, the fixed
760x660 window, and the settingsTab scroll/padding wrapper.

New per-tab views (Views/Settings/): General, Provider, ApiUsage, Cost,
About — each carrying current-master content. Account sheets and rows
(AddClaudeAccountSheet, AddCodexAccountSheet, AccountProfileRow,
CodexAccountProfileRow) extracted to their own files.

Reusable primitives moved to Views/Components/SettingsComponents.swift
(private->internal), including a new ExtraUsageRow that dedupes the two
extraUsageRow call sites.

ProviderSettingsFacts value type centralizes the repeated 4-way
`switch service` fact derivation (source/status/plan/statusColor derived
once); every switch handles .grok. Provider usage rows use the unified
LimitRow(density:.regular).

Pure refactor — no behavior/visual change. Verified: swift build clean,
swiftlint --strict clean, SettingsViewSmokeTests + new
ProviderSettingsFactsTests green, and all 6 settings tabs walked in the
running app render identically.
