+++ .agents/SESSIONS/2026-07-14.md
@@ -975,3 +975,52 @@ losing the only decision onboarding asks for (launch-at-login).
 - `testEnableLaunchAtLoginRegistersAndCompletesOnboarding` /
   `testNotNowCompletesWithoutRegistering` — extended to assert the explicit choice
   persists across a reload (`hasCompletedFirstRun` true, reloaded store silent).
+
+---
+
+# Session: SettingsView de-monolith (redo of deferred #192)
+
+## Context
+Draft PR #192 ("de-monolith SettingsView 2052→85") was deferred/held: its split
+files were computed against the *original* monolith and predated 6 merged PRs
+(#187 chip badges, #188 glass CTAs, #190 a11y/Dynamic Type, #191 tokens, #193
+EmptyStateCard notices, #196 Grok). Merging it would have regressed the settings
+UI. This session re-does the split against **current master**, using #192's
+branch (`claude/objective-aryabhata-cea348`) only as the structural blueprint.
+
+## What changed
+Split `MeterBar/Views/SettingsView.swift` (2273 → 88 lines) into a thin shell
+(TabView + 760×660 frame + `settingsTab` wrapper only). New files:
+- `Views/Components/SettingsComponents.swift` — shared primitives
+  (`SettingsPanelSection`, `SettingsRowView` w/ a11y `.combine`, `SettingsInfoRow`,
+  `SettingsNotice`, `SettingsDivider`, `StatusPill`, `AdminKeySettingsRow`,
+  `SettingsReadonlyField`, input modifiers) + new `ExtraUsageRow` (dedupes the
+  two `extraUsageRow` call sites). `private`→`internal`.
+- `Views/Settings/ProviderSettingsFacts.swift` — value type centralizing the
+  repeated 4-way `switch service` fact derivation (source/status/plan/statusColor
+  derived once). Every switch handles `.grok`. Houses `QuotaBand.severity` +
+  `String.nilIfEmpty`.
+- `Views/Settings/{General,Provider,ApiUsage,Cost,About}SettingsView.swift` —
+  per-tab views carrying **current-master** content (chip badges, `.glass`/
+  `.glassProminent` CTAs, `EmptyStateCard` notices, `LimitRow(density:.regular)`,
+  token spacing, VoiceOver labels, Grok, menuBarDisplaySection).
+- `Views/Settings/{AddClaudeAccountSheet,AddCodexAccountSheet,ClaudeAccountProfileRow,CodexAccountProfileRow}.swift`
+  — extracted verbatim.
+
+Pure refactor: no behavior/visual change. Xcode uses synchronized folders and
+SwiftPM globs `MeterBar/`, so no project-file edits were needed.
+
+## Verification
+- `swift build` clean; `swiftlint --strict` 0 violations on all new/changed files.
+- `swift test --filter SettingsViewSmokeTests` green (3/3) +
+  new `ProviderSettingsFactsTests` (14/14, covers every service incl `.grok`).
+- Full suite: only the 13 pre-existing `SessionWake{Agent,Controller}Tests`
+  failures (filesystem-sandbox lock-dir denial) — confirmed identical on a
+  stashed clean master; unrelated to this View-layer refactor.
+- Ran the real app (`MeterBar Dev.app`, ad-hoc signed) and walked all 6 settings
+  tabs: General (Grok toggle + Menu Bar section), Providers (facts Overview +
+  LimitRow usage + Grok pane), API Usage, Cost (glassProminent scan), About —
+  UI identical to before.
+
+## Follow-up
+PR supersedes draft #192; asks reviewer to close #192.
